pyproject.toml
[project]
name = "syncthing-monitor"
version = "0.1.0"
description = "Monitor syncthing status and send notifications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
